+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.1
 
 * Add AndroidContentProvider example and make README setup instructions more clear
+* Avoid breaking changes in Android StandardMessageCodec Flutter 2.13.0 by converting AndroidContentProviderMessageCodec from Kotlin to Java
 
 ## 0.2.0